Refactor: Result Charts
-extensive refactor
-rename variables / functions / types
-decouple Result Chart from History Chart
-match line color
-match tooltips
-localize types / remove file
