--- packages/app/src/types/result.ts
@@ -1,15 +1,0 @@
-import { ParsedRacesResult } from "@/app/result/loaders";
-
-export type ResultChartLine = {
-  dataKey: keyof ParsedRacesResult;
-  stroke: string;
-};
-
-export interface ResultCardProps {
-  title: string;
-  value: string | undefined;
-}
-
-export interface ResultsChartProps {
-  searchParams: { snippetId: string };
-}
